refactor(blog): reorganise blog models and helpers for improved structure
